+++ docs/source/history.md
@@ -213,5 +213,7 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 - Examples:
   - New in `foxes.opt`: Example `layout_regular_grid`, demonstrates regular grid layout optimization
   - New in `foxes.opt`: Example `geom_reggrids`, purely geometrical regular layout optimization
+- Utils:
+  - New functions for shape file handling: `read_shp`, `shp2csv`, `read_shp_polygons`
   
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.3.1](https://github.com/FraunhoferIWES/foxes/commits/v0.3.1)